requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==26.0.0
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tenacity==8.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
